requirements.txt
asgiref==3.5.2
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
psycopg==3.1.4
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
tzdata==2022.6

requests~=2.28.1